requirements.txt
fitbit==0.3.1
Flask==3.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Werkzeug==3.0.1

